spring boot crud

+++ src/main/java/com/mycompany/ecom/Dto/ProductDto.java
@@ -1,15 +1,19 @@
 package com.mycompany.ecom.Dto;
 
 import lombok.Data;
-import org.springframework.stereotype.Component;
 
 import java.time.LocalDate;
 
 @Data
 public class ProductDto {
-    private int id;
+
+    private Long id;
+
     private String name;
+
     private String description;
+
     private Double price;
+
     private LocalDate manufactureDate;
 }
